requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
python-multipart
python-dotenv
google-adk>=0.1.6
litellm>=1.25.0
google-generativeai>=0.8.5
